requirements.txt
skorch==0.8 # use the correct skorch version if you want to use the pretrained models
torchvision
matplotlib
tqdm
scikit-learn
joblib
h5py
seaborn
scikit-image